+++ src/main/java/duke/command/Commands.java
@@ -17,10 +17,21 @@ public enum Commands {
         this.command = command;
     }
 
+    /**
+     * Checks if the first word of the user input matches this command.
+     *
+     * @param firstWord first word of user input
+     * @return true if the first word of the user input matches this command
+     */
     public boolean isCommand(String firstWord) {
         return firstWord.equals(command.getCommandString());
     }
 
+    /**
+     * Returns the command of this type
+     *
+     * @return the Command
+     */
     public Command getCommand() {
         return command;
     }
