+++ src/main/java/duke/command/ExitCommand.java
@@ -1,15 +1,20 @@
 package duke.command;
 
 import duke.Duke;
+import duke.exception.DukeException;
 
 public class ExitCommand extends Command {
     public ExitCommand() {
         setCommandString("bye");
     }
 
+    /**
+     * Exits the application.
+     *
+     * @param input Full user input
+     */
     @Override
     public void parse(String input) {
-        // String data = input.substring(getCommandLength());
         Duke.exit();
     }
 }
