+++ src/main/java/duke/io/TextColor.java
@@ -11,6 +11,11 @@ public enum TextColor {
         this.ansiCode = ansiCode;
     }
 
+    /**
+     * The ANSI code to set text color to this color
+     *
+     * @return the ANSI code, in String form, to set text color
+     */
     @Override
     public String toString() {
         return ansiCode;
