+++ src/main/java/duke/task/TaskList.java
@@ -18,6 +18,13 @@ public class TaskList {
         this.taskList = new ArrayList<>();
     }
 
+    /**
+     * Adds the task into the taskList
+     * Also saves the taskList and queues a message to be printed
+     *
+     * @param task The task to be added
+     * @throws DukeException I/O exception from saving the taskList state
+     */
     public void addTask(Task task) throws DukeException {
         taskList.add(task);
         save();
@@ -26,6 +33,13 @@ public class TaskList {
                 task.toString() + "\n" + taskLengthReport(), TextColor.DEFAULT);
     }
 
+    /**
+     * Marks the specified task as done
+     * Also saves the taskList and queues a message to be printed
+     *
+     * @param taskNum The number of the task to be marked as done (1-indexed)
+     * @throws DukeException Exception from out of range index, or from saving the list
+     */
     public void doTask(int taskNum) throws DukeException {
         int idx = getTaskIndexFromTaskNum(taskNum);
         Task task = taskList.get(idx);
@@ -36,6 +50,13 @@ public class TaskList {
                 task.toString(), TextColor.DEFAULT);
     }
 
+    /**
+     * Deletes the specified task
+     * Also saves the taskList and queues a message to be printed
+     *
+     * @param taskNum The number of the task to be deleted (1-indexed)
+     * @throws DukeException Exception from out of range index, or from saving the list
+     */
     public void deleteTask(int taskNum) throws DukeException {
         int idx = getTaskIndexFromTaskNum(taskNum);
 
@@ -47,6 +68,12 @@ public class TaskList {
                 task.toString() + "\n" + taskLengthReport(), TextColor.DEFAULT);
     }
 
+    /**
+     * Deletes all tasks that have been marked as done
+     * Also saves the taskList and queues a message to be printed
+     *
+     * @throws DukeException I/O exception from saving the taskList state
+     */
     public void deleteDone() throws DukeException {
         taskList.removeIf(Task::isDone);
         save();
@@ -54,6 +81,12 @@ public class TaskList {
                 taskLengthReport(), TextColor.DEFAULT);
     }
 
+    /**
+     * Deletes all tasks that are expired
+     * Also saves the taskList and queues a message to be printed
+     *
+     * @throws DukeException I/O exception from saving the taskList state
+     */
     public void deleteExpired() throws DukeException {
         taskList.removeIf(Task::isExpired);
         save();
@@ -61,6 +94,9 @@ public class TaskList {
                 taskLengthReport(), TextColor.DEFAULT);
     }
 
+    /**
+     * Queues a message listing all the tasks to be printed
+     */
     public void list() {
         int size = taskList.size();
         if (size == 0) {
@@ -74,7 +110,13 @@ public class TaskList {
         }
     }
 
-    // parses the user's input and returns the index of the task in question
+    /**
+     * Returns the 0-indexed index of the task from it's 1-indexed index
+     *
+     * @param taskNum 1-indexed index of the task
+     * @return 0-indexed index of the task
+     * @throws DukeException Exception thrown if the index is out of range
+     */
     private int getTaskIndexFromTaskNum(int taskNum) throws DukeException {
         if (taskNum <= 0 || taskNum > taskList.size()) {
             throw new DukeException("Please input the ID of a task!");
@@ -84,16 +126,30 @@ public class TaskList {
         return taskNum - 1;
     }
 
-    // returns a string telling the user how many tasks are in the list
+    /**
+     * Returns a string telling the user how many tasks are in the list
+     *
+     * @return string telling the user how many tasks are in the list
+     */
     private String taskLengthReport() {
         return "Now you have " + taskList.size()
                 + (taskList.size() != 1 ? " tasks" : " task") + " in the list.";
     }
 
+    /**
+     * Saves the state of the taskList
+     *
+     * @throws DukeException I/O exception from saving
+     */
     private void save() throws DukeException {
         Duke.storage.save(convertToSaveString());
     }
 
+    /**
+     * Returns a string representing the tasks compliant to the save format
+     *
+     * @return a string representing the tasks compliant to the save format
+     */
     private String convertToSaveString() {
         StringBuilder sb = new StringBuilder();
 
@@ -112,7 +168,11 @@ public class TaskList {
         return sb.toString();
     }
 
-    // for testing
+    /**
+     * Returns the taskList to operate on without any side effects
+     *
+     * @return the taskList
+     */
     protected ArrayList<Task> getList() {
         return taskList;
     }
