+++ src/main/java/duke/task/ToDo.java
@@ -7,6 +7,12 @@ public class ToDo extends Task {
         super(name);
     }
 
+    /**
+     * Loads a ToDo from data parsed from the save file
+     *
+     * @param loadData A line from the csv, split by commas
+     * @return ToDo created from provided data
+     */
     public static ToDo load(String[] loadData) {
         boolean done = loadData[1].equals("o");
         String name = loadData[2];
@@ -19,22 +25,21 @@ public class ToDo extends Task {
         return todo;
     }
 
-    @Override
-    public boolean isExpired() {
-        return false;
-    }
-
-    @Override
-    public TextColor getListColor() {
-        return TextColor.DEFAULT;
-    }
-
-    // prints in red if the deadline is expired, yellow if deadline is today
+    /**
+     * Returns a string representation of the ToDo and its status
+     *
+     * @return string representation of the ToDo
+     */
     @Override
     public String toString() {
         return "[T] " + super.toString();
     }
 
+    /**
+     * Returns a string representing the ToDo compliant to the saveFile format
+     *
+     * @return String to be saved as a line in save.csv
+     */
     @Override
     public String getSaveString() {
         return "t," + super.getSaveString();
